gui_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
bzip2=1.0.6=h14c3975_5
ca-certificates=2018.03.07=0
cairo=1.14.12=h8948797_3
certifi=2018.8.24=py37_1
cffi=1.11.5=py37he75722e_1
dbus=1.13.2=h714fa37_1
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
graphite2=1.3.12=h23475e2_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
intel-openmp=2019.0=118
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.2.1=hb9ed12e_0
libpng=1.6.34=hb9fc6fc_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
mkl=2019.0=118
mkl_fft=1.0.4=py37h4414c95_1
mkl_random=1.0.1=py37h4414c95_1
ncurses=6.1=hf484d3e_0
ninja=1.8.2=py37h6bb024c_1
numpy=1.15.1=py37h1d66e8a_0
numpy-base=1.15.1=py37h81de0dd_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.0.2p=h14c3975_0
pandas=0.23.4=py37h04863e7_0
pcre=8.42=h439df22_0
pillow=5.2.0=py37heded4f4_0
pip=10.0.1=py37_0
pixman=0.34.0=hceecf20_3
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.18=py37_1
pyqt=5.9.2=py37h05f1152_2
python=3.7.0=hc3d631a_0
python-dateutil=2.7.3=py37_0
pytorch=0.4.1=py37_py36_py35_py27__9.0.176_7.1.2_2
pytz=2018.5=py37_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
scipy=1.1.0=py37hfa4b5c9_1
setuptools=40.2.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.11.0=py37_1
sqlite=3.24.0=h84994c4_0
tk=8.6.8=hbc83047_0
tqdm=4.26.0=py37h28b3542_0
wheel=0.31.1=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
